Move new classes to Helpers
